refactor: move runtime types to where they belong
Storage domain types (ConversationMeta, EventLine, ArchiveSegment,
sender_slug) move into core::storage — they describe persistence
format, not runtime behavior.

Runtime abstractions (Config, Hook, Conversation) move from core into
the runtime crate — they're only useful with a runtime. Core no longer
has a runtime module; the runtime crate owns all runtime abstractions.
