+++ crates/core/src/test_utils/mem.rs
@@ -3,8 +3,10 @@
 use crate::{
     AgentConfig, AgentId, ManifestConfig, NodeConfig,
     model::HistoryEntry,
-    runtime::conversation::{ArchiveSegment, ConversationMeta, EventLine},
-    storage::{MemoryEntry, SessionHandle, SessionSnapshot, SessionSummary, Skill, Storage},
+    storage::{
+        ArchiveSegment, ConversationMeta, EventLine, MemoryEntry, SessionHandle, SessionSnapshot,
+        SessionSummary, Skill, Storage,
+    },
 };
 use anyhow::Result;
 use std::{collections::HashMap, sync::Mutex};
